+++ GOtravel/GoTravel/View/ViewController/Home/TripCell.swift
